FIX various fixes - size_t vs. int mismatches - few incorrect == vs = - few missing return values - few incorrect logging calls

+++ src/asset/ModelMaterialInit.cpp
@@ -21,7 +21,7 @@ void ModelMaterialInit::prepareVertices(
     // https://paroj.github.io/gltut/Basic%20Optimization.html
     {
         const bool single = materialVBO.m_materials.size() == 1;
-        const int count = single ? 1 : vertices.size();
+        const size_t count = single ? 1 : vertices.size();
         auto& entries = materialVBO.m_entries;
         entries.reserve(count);
